backend/requirements.txt
fastapi==0.104.1
fastapi-socketio==0.0.10
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-socketio==5.10.0
python-engineio==4.7.1